requirements.txt
browser-use
playwright
langchain-openai
python-dotenv